# server_source/reports/openapi_manual_export.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Murim Inn Rebuild Server",
    "version": "0.4.0"
  },
  "paths": {
...
    "/admin/tools/grant": {
      "post": {
        "tags": [
          "AdminTools",
          "LiveOps"
        ],
        "summary": "Grant Currency",
        "operationId": "grant_currency_admin_tools_grant_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrantCurrencyReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/tools/ban": {
      "post": {
        "tags": [
          "AdminTools",
          "LiveOps"
        ],
        "summary": "Ban Account",
        "operationId": "ban_account_admin_tools_ban_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BanReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/tools/unban": {
      "post": {
        "tags": [
          "AdminTools",
          "LiveOps"
        ],
        "summary": "Unban Account",
        "operationId": "unban_account_admin_tools_unban_post",
        "parameters": [
          {
            "name": "accountId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Accountid"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/tools/ban/{account_id}": {
      "get": {
        "tags": [
          "AdminTools",
          "LiveOps"
        ],
        "summary": "Get Ban Status",
        "operationId": "get_ban_status_admin_tools_ban__account_id__get",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Account Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/tools/mailbox/grant": {
      "post": {
        "tags": [
          "AdminTools",
          "LiveOps"
        ],
        "summary": "Admin Mailbox Grant",
        "operationId": "admin_mailbox_grant_admin_tools_mailbox_grant_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MailGrantReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/tools/mailbox/{account_id}": {
      "get": {
        "tags": [
          "AdminTools",
          "LiveOps"
        ],
        "summary": "Admin List Mailbox",
        "operationId": "admin_list_mailbox_admin_tools_mailbox__account_id__get",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Account Id"
            }
          },
          {
            "name": "includeClaimed",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Includeclaimed"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/tools/mailbox/{message_id}/claim": {
      "post": {
        "tags": [
          "AdminTools",
          "LiveOps"
        ],
        "summary": "Admin Claim Mailbox",
        "operationId": "admin_claim_mailbox_admin_tools_mailbox__message_id__claim_post",
        "parameters": [
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Message Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/tools/account-summary/{account_id}": {
      "get": {
        "tags": [
          "AdminTools",
          "LiveOps"
        ],
        "summary": "Admin Account Summary",
        "operationId": "admin_account_summary_admin_tools_account_summary__account_id__get",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Account Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/tools/announcement/upsert": {
      "post": {
        "tags": [
          "AdminTools",
          "LiveOps"
        ],
        "summary": "Admin Upsert Announcement",
        "operationId": "admin_upsert_announcement_admin_tools_announcement_upsert_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnnouncementUpsertReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnnouncementUpsertReq": {
        "properties": {
          "announcementId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Announcementid"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "startsAtEpochSec": {
            "type": "integer",
            "title": "Startsatepochsec"
          },
          "endsAtEpochSec": {
            "type": "integer",
            "title": "Endsatepochsec"
          },
          "priority": {
            "type": "integer",
            "title": "Priority",
            "default": 100
          },
          "isEnabled": {
            "type": "boolean",
            "title": "Isenabled",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "title",
          "body",
          "startsAtEpochSec",
          "endsAtEpochSec"
        ],
        "title": "AnnouncementUpsertReq"
      },
...
      "BanReq": {
        "properties": {
          "accountId": {
            "type": "string",
            "title": "Accountid"
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "default": ""
          },
          "durationSec": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Durationsec",
            "description": "If omitted -> permanent. If provided -> seconds from now."
          }
        },
        "type": "object",
        "required": [
          "accountId"
        ],
        "title": "BanReq"
      },
...
      "GrantCurrencyReq": {
        "properties": {
          "accountId": {
            "type": "string",
            "title": "Accountid"
          },
          "currency": {
            "type": "string",
            "enum": [
              "gold",
              "gems",
              "exp"
            ],
            "title": "Currency"
          },
          "amount": {
            "type": "integer",
            "title": "Amount",
            "description": "can be negative for rollback; must be non-zero"
          },
          "hunterId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hunterid",
            "description": "Optional. If provided, directly mutates the hunter snapshot for convenience."
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "default": ""
          },
          "idempotencyKey": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotencykey",
            "description": "Optional. If omitted, uses Idempotency-Key header or auto-generates (not idempotent)."
          }
        },
        "type": "object",
        "required": [
          "accountId",
          "currency",
          "amount"
        ],
        "title": "GrantCurrencyReq"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MailGrantReq": {
        "properties": {
          "accountId": {
            "type": "string",
            "title": "Accountid"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "rewardCurrency": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "gold",
                  "gems",
                  "exp"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Rewardcurrency"
          },
          "rewardAmount": {
            "type": "integer",
            "title": "Rewardamount",
            "default": 0
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "default": ""
          },
          "sourceId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourceid"
          }
        },
        "type": "object",
        "required": [
          "accountId",
          "title",
          "body"
        ],
        "title": "MailGrantReq"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}